Switch to a more sophisticated Makefile.
This is to make it easier to grow the library without copy-pasting the
Makefile, and to make it easier to build the library for e.g. AVR instead
of local architecture.

So far I've been unable to get the conditional MCU argument to work right
so that part doesn't work yet.